Added kernel module, debug, and headers when perf is enabled
Many perf tools require linux headers and other
related kernel-specific files to operate.

Instead of requiring every container to have these packages,
bind-mounting them is more convenient, and ensures that they
are always accurate and correct for the running kernel
as it changes.

Additionally it works for any container, regardless of their distro.